chore: reformat project with new line length

+++ src/services/favourites.ts
@@ -1,26 +1,18 @@
 import type { Database } from '~/supabase.types'
 import type { Favourite } from '~/types'
 
-export async function getUserHasFavourites(
-  userId: string | undefined
-): Promise<Favourite[]> {
+export async function getUserHasFavourites(userId: string | undefined): Promise<Favourite[]> {
   const client = useSupabaseClient<Database>()
   if (!userId) return []
 
-  const { data, error } = await client
-    .from('favourites')
-    .select()
-    .eq('user_id', userId)
+  const { data, error } = await client.from('favourites').select().eq('user_id', userId)
 
   if (error) throw new Error(error.message)
 
   return data as Favourite[]
 }
 
-export async function addFavourite(
-  noteId: number,
-  userId: string
-): Promise<Favourite> {
+export async function addFavourite(noteId: number, userId: string): Promise<Favourite> {
   const client = useSupabaseClient<Database>()
 
   const { data, error } = await client
@@ -37,10 +29,7 @@ export async function addFavourite(
 export async function removeFavourite(noteId: number, userId: string) {
   const client = useSupabaseClient<Database>()
 
-  const { error, status } = await client
-    .from('favourites')
-    .delete()
-    .match({ user_id: userId, note_id: noteId })
+  const { error, status } = await client.from('favourites').delete().match({ user_id: userId, note_id: noteId })
 
   if (error) throw new Error(error.message)
